rename homepage-radio component
