+++ src/app/app.module.ts
@@ -10,12 +10,13 @@ import { HomeComponent } from './homepage/home/home.component';
 import { ProfilePicComponent } from './profile-pic/profile-pic.component';
 import { ContactLinksComponent } from './homepage/contact-links/contact-links.component';
 import { AboutComponent } from './homepage/about/about.component';
-import { RadioComponent } from './homepage-radio/radio/radio.component';
+import { RadioComponent } from './homepage-radio/homepage-radio.component';
 import { SidebarLinksComponent } from './sidebar-links/sidebar-links.component';
 import { UrlSerializer } from '@angular/router';
 import { LowerCaseUrlSerializer } from './lower-case-url-serializer';
 import { SidebarImgLinksComponent } from './sidebar-img-links/sidebar-img-links.component';
 import { IntroComponent } from './homepage-radio/intro/intro.component';
+import { FeaturedLinksComponent } from './homepage-radio/featured-links/featured-links.component';
 
 @NgModule({
   declarations: [
@@ -27,7 +28,8 @@ import { IntroComponent } from './homepage-radio/intro/intro.component';
     RadioComponent,
     SidebarLinksComponent,
     SidebarImgLinksComponent,
-    IntroComponent
+    IntroComponent,
+    FeaturedLinksComponent
   ],
   imports: [BrowserModule, FontAwesomeModule, NgbModule, LightboxModule, AppRoutingModule],
   providers: [
