+++ src/app/homepage-radio/homepage-radio.component.html
@@ -21,6 +21,9 @@
       <div class="row no-gutters">
         <app-intro></app-intro>
       </div>
+      <div class="row no-gutters">
+        <app-featured-links></app-featured-links>
+      </div>
     </div>
   </div>
 </div>
