+++ src/app/homepage-radio/homepage-radio.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { RadioComponent } from './radio.component';
+import { RadioComponent } from './homepage-radio.component';
 
 describe('RadioComponent', () => {
   let component: RadioComponent;
@@ -8,9 +8,8 @@ describe('RadioComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ RadioComponent ]
-    })
-    .compileComponents();
+      declarations: [RadioComponent]
+    }).compileComponents();
   }));
 
   beforeEach(() => {
